fix: more tweaks to expose editor cache
