+++ Scripts/Editor/CollectableScriptableObjectPropertyDrawer.cs
@@ -111,7 +111,7 @@ namespace BrunoMikoski.ScriptableObjectCollections
                         EditorGUI.indentLevel++;
                         using (new EditorGUILayout.VerticalScope("Box"))
                         {
-                            Editor editor = CollectionUtility.GetOrCreateEditorForItem(collectableItem);
+                            Editor editor = EditorsCache.GetOrCreateEditorForItem(collectableItem);
                             using (EditorGUI.ChangeCheckScope changedCheck = new EditorGUI.ChangeCheckScope())
                             {
                                 GUILayout.Space(10);
